server/python/requirements.txt
CherryPy==18.6.0
gunicorn==20.1.0
supervisor==4.2.2
packaging
numpy
wave
jiwer
webrtcvad
transformers
tqdm==4.61.0
pyyaml==5.4.1
torchaudio==0.7.2
torch==1.7.1
